WIP:
Doesn't this really disable verification as such?  Not just fetching?
Maybe only the fetching could be suppressed so that cached CRLs (and even
OCSP responses) could still be used?  Strangely, there is already an
`online` flag that has to be TRUE so the plugin does anything at all.

Also, if both methods are disabled the revocation plugin now fails with
CRED_HOOK_VALIDATION_FAILED. Is that intended?  Otherwise, that last
call_hook() call should probably be put in e.g. an
`if (this->enable_ocsp || this->enable_crl)` block.